# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Go + Gin baseline for a learning-focused streaming system.",
        "title": "demo-streaming API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/stream-publish-sessions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stream-publish-sessions"
                ],
                "summary": "List stream publish sessions",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page number (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page size (default 20, max 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/streampublishsessions.ListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/streampublishsessions.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/streampublishsessions.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/streampublishsessions.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stream-publish-sessions"
                ],
                "summary": "Create stream publish session",
                "parameters": [
                    {
                        "description": "create publish session",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/streampublishsessions.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/streampublishsessions.CreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/streampublishsessions.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/streampublishsessions.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/streampublishsessions.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/streampublishsessions.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/stream-publish-sessions/stop-all": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stream-publish-sessions"
                ],
                "summary": "Stop all live stream publish sessions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/streampublishsessions.StopAllResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/streampublishsessions.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/streampublishsessions.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/stream-publish-sessions/{id}/start": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stream-publish-sessions"
                ],
                "summary": "Start stream publish session",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/streampublishsessions.StartResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/streampublishsessions.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/streampublishsessions.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/streampublishsessions.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/streampublishsessions.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/streampublishsessions.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/streampublishsessions.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/stream-publish-sessions/{id}/stop": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stream-publish-sessions"
                ],
                "summary": "Stop stream publish session",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/streampublishsessions.StopResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/streampublishsessions.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/streampublishsessions.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/streampublishsessions.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/streampublishsessions.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/streampublishsessions.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/streampublishsessions.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "streampublishsessions.CreateIngestResponse": {
            "type": "object",
            "properties": {
                "provider": {
                    "type": "string",
                    "example": "srs"
                },
                "rtmp_url": {
                    "type": "string",
                    "example": "rtmp://localhost:1935/live"
                },
                "stream_key": {
                    "type": "string"
                }
            }
        },
        "streampublishsessions.CreateRequest": {
            "type": "object",
            "properties": {
                "title": {
                    "type": "string"
                }
            }
        },
        "streampublishsessions.CreateResponse": {
            "type": "object",
            "properties": {
                "ingest": {
                    "$ref": "#/definitions/streampublishsessions.CreateIngestResponse"
                },
                "playback_url_cdn": {
                    "type": "string"
                },
                "session_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "streampublishsessions.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "streampublishsessions.ListItemResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "integer"
                },
                "ended_at": {
                    "type": "integer"
                },
                "playback_id": {
                    "type": "string"
                },
                "playback_url_cdn": {
                    "type": "string"
                },
                "renditions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/streampublishsessions.ListRenditionResponse"
                    }
                },
                "session_id": {
                    "type": "integer"
                },
                "started_at": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "streampublishsessions.ListRenditionResponse": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "playlist_path": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "streampublishsessions.ListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/streampublishsessions.ListItemResponse"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "streampublishsessions.StartResponse": {
            "type": "object",
            "properties": {
                "session_id": {
                    "type": "integer"
                },
                "started_at": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "streampublishsessions.StopAllResponse": {
            "type": "object",
            "properties": {
                "ended_at": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "stopped_session_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "streampublishsessions.StopResponse": {
            "type": "object",
            "properties": {
                "ended_at": {
                    "type": "integer"
                },
                "session_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}